requirements.txt
openai==1.109.1
python-dotenv==1.1.1
langchain-openai==0.3.16
numpy==1.26.4
pyyaml>=6.0

